Trait storage_proofs_core::compound_proof::CircuitComponent

pub trait CircuitComponent {
    type ComponentPrivateInputs: Default + Clone;
}

CircuitComponent exists so parent components can pass private inputs to their subcomponents when calling CompoundProof::circuit directly. In general, there are no internal private inputs, and a default value will be passed. CompoundProof::circuit implementations should exhibit default behavior when passed a default ComponentPrivateinputs.

Required Associated Types

type ComponentPrivateInputs: Default + Clone