Module storage_proofs_core::merkle

Structs

DiskStore

The Disk-only store is used to reduce memory to the minimum at the cost of build time performance. Most of its I/O logic is in the store_copy_from_slice and store_read_range functions.

ExternalReader

InclusionPath

MerkleProof

Representation of a merkle proof.

MerkleTreeWrapper

PathElement

Traits

MerkleProofTrait

Trait to abstract over the concept of Merkle Proof.

MerkleTreeTrait

Trait used to abstract over the way Merkle Trees are constructed and stored.

Store

Backing store of the merkle tree.

Functions

base_path_length

compound_path_length

compound_tree_height

create_base_lcmerkle_tree

Construct a new level cache merkle tree, given the specified config.

create_base_merkle_tree

create_disk_tree

create_lc_tree

create_tree

generate_tree

Only used for testing, but can’t cfg-test it as that stops exports.

get_base_tree_count

get_base_tree_leafs

split_config

split_config_and_replica

split_config_wrapped

Type Definitions

BinaryMerkleTree

A binary merkle tree, where all levels have arity 2. It’s fully persisted to disk.

DiskTree

A tree that is fully persisted to disk.

LCTree

A tree that is partially stored on disk, some levels are in memory.

ResTree