Struct storage_proofs_core::merkle::DiskStore

pub struct DiskStore<E>where
    E: Element,{ /* private fields */ }

The Disk-only store is used to reduce memory to the minimum at the cost of build time performance. Most of its I/O logic is in the store_copy_from_slice and store_read_range functions.

Implementations

impl<E> DiskStore<E>where E: Element,

pub fn new_from_disk_with_path<P>( size: usize, data_path: P ) -> Result<DiskStore<E>, Error>where P: AsRef<Path>,

pub fn is_consistent( store_range: usize, _branches: usize, config: &StoreConfig ) -> Result<bool, Error>

pub fn store_size(&self) -> usize

pub fn store_read_range( &self, start: usize, end: usize ) -> Result<Vec<u8, Global>, Error>

pub fn store_read_into( &self, start: usize, end: usize, buf: &mut [u8] ) -> Result<(), Error>

pub fn store_copy_from_slice( &mut self, start: usize, slice: &[u8] ) -> Result<(), Error>