Type Alias storage_proofs_core::merkle::DiskTree

pub type DiskTree<H, U, V, W> = MerkleTreeWrapper<H, DiskStore<<H as Hasher>::Domain>, U, V, W>;

A tree that is fully persisted to disk.

It’s generic over the hash function H, the base arity U, sub-tree arity V and top-tree arity W.

The base arity is used for for all levels up to the top. Non zero arties of the top-tree and/or sub-tree each add another layer on top. So a tree with e.g. U = 8, V = 4, W = 2 would create a tree where the top level has two children, the levels below 4 children and all other levels below have 8 children.

Aliased Type

struct DiskTree<H, U, V, W> {
    pub inner: MerkleTree<<H as Hasher>::Domain, <H as Hasher>::Function, DiskStore<<H as Hasher>::Domain>, U, V, W>,
    pub h: PhantomData<H>,
}

Fields
inner: MerkleTree<<H as Hasher>::Domain, <H as Hasher>::Function, DiskStore<<H as Hasher>::Domain>, U, V, W>h: PhantomData<H>